model/requirements.txt
# Core ML and API dependencies
tensorflow==2.13.0
flask==2.3.3
opencv-python==4.8.1.78
pillow==10.0.1
numpy==1.24.3
requests==2.31.0

# Image processing
imageio==2.31.3
scikit-image==0.21.0

# API and web utilities
flask-cors==4.0.0
werkzeug==2.3.7
gunicorn==21.2.0

# Logging and monitoring
python-dotenv==1.0.0
pydantic==2.3.0

# Optional: For production deployment
waitress==2.1.2
psutil==5.9.5

# For base64 and URL handling
urllib3==2.0.4
certifi==2023.7.22